Changed injector to reflective injection
DLL injections are now being made using Stephen Fewer's "Reflective DLL
Injection" method, thus reducing the number of disk operations and
becoming much harder to detect. The method also allowing injecting to
both x86 and x64 processes from a x64 process. This also solved the need
for delay, but the code still stays there for in case it'll be needed in
the future.